Implemented CheckOut method

+++ PromotionEngine/PromotionEngine/Module/Promotion.cs
@@ -5,9 +5,9 @@ namespace PromotionEngine.Module
     public class Promotion
     {
         public Dictionary<string, int> ProductInfoDict { get; set; }
-        public decimal PromoPrice { get; set; }
+        public int PromoPrice { get; set; }
 
-        public Promotion(Dictionary<string, int> prodInfoDict, decimal promoPrice)
+        public Promotion(Dictionary<string, int> prodInfoDict, int promoPrice)
         {
             ProductInfoDict = prodInfoDict;
             PromoPrice = promoPrice;
